feat: deprecate unused fields in BaseActivity
The commit message summarizes the changes made in the diffs. The diffs show that two fields, `context` and `queue`, in the `BaseActivity` class have been marked as deprecated using the `@Deprecated` annotation. This indicates that these fields are no longer recommended for use and may be removed in future versions of the code.